+++ projects/plugins/boost/.prettierrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+	...require( '@wordpress/prettier-config' ),
+	...require( '../../../.prettierrc'),
+	plugins: [ 'prettier-plugin-svelte' ],
+	svelteStrictMode: false,
+	svelteBracketNewLine: true,
+	svelteIndentScriptAndStyle: true,
+};