+++ projects/plugins/jetpack/_inc/client/search/dashboard/instant-search-upsell-nudge.jsx
@@ -9,17 +9,19 @@ import { __ } from '@wordpress/i18n';
  */
 import './instant-search-upsell-nudge.scss';
 
-const InstantSearchUpsellNudge = props => {
+const InstantSearchUpsellNudge = ( props = { upgrade: true } ) => {
 	return (
-		<a className="jp-instant-search-upsell-nudge" href={ props.href }>
+		<a className="jp-instant-search-upsell-nudge jp-search-dashboard-cut" href={ props.href }>
 			<span>
 				{ __(
 					'Offer instant search results to your visitors as soon as they start typing. ',
 					'jetpack'
 				) }
-				<b>{ __( 'Upgrade to Jetpack Instant Search now', 'jetpack' ) }</b>
 			</span>
-			<div className="jp-instant-search-upsell-nudge__button-arrow">&rarr;</div>
+			<span>
+				{ props.upgrade && <b>{ __( 'Upgrade to Jetpack Instant Search now', 'jetpack' ) }</b> }
+				{ ! props.upgrade && <b>{ __( 'Purchase Jetpack Instant Search now', 'jetpack' ) }</b> }
+			</span>
 		</a>
 	);
 };